requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
httpx==0.28.1
python-multipart==0.0.20
python-dotenv==1.0.1
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
PyYAML==6.0.1
markdown==3.5.2
qrcode==7.4.2
pillow==10.2.0
pypdf==5.1.0
python-docx==1.1.2
alibabacloud-dysmsapi20170525==4.5.0
